refactor: extracted session controller logic into helper methods
